+++ src/blocks/gravity-form-renderer/_partials/fields/name.php
@@ -20,7 +20,7 @@
 			<div class="col-auto flex-grow-1" data-wp-class--d-none="callbacks.isHiddenInput">
 				<div class="invalid-feedback" data-wp-bind--id="callbacks.nameInputErrorId" data-wp-bind--hidden="!callbacks.hasNameInputError" data-wp-text="callbacks.nameInputError"></div>
 				<input class="form-control" type="text" data-wp-bind--id="callbacks.nameInputId" data-wp-bind--name="callbacks.nameInputName"
-					data-wp-bind--required="context.field.isRequired" data-wp-bind--value="callbacks.nameInputValue" data-wp-bind--autocomplete="callbacks.nameInputAutocomplete"
+					data-wp-bind--required="context.field.isRequired" data-wp-bind--value="callbacks.prefilledNameInputValue" data-wp-bind--autocomplete="callbacks.nameInputAutocomplete"
 					data-wp-bind--aria-invalid="callbacks.hasNameInputError" data-wp-bind--aria-describedby="callbacks.nameInputErrorId" data-wp-on--input="actions.updateNameInputValue"
 					data-wp-class--is-invalid="callbacks.hasNameInputError" />
 				<label class="form-label has-root-font-size" data-wp-bind--for="callbacks.nameInputId">