+++ src/blocks/gravity-form-renderer/_store/types.ts
@@ -1,6 +1,9 @@
 export type GravityFormsContext = {
 	formId: number;
 	form: any | null;
+	field?: any;
+	input?: any;
+	prefilledValues: Record< string, string >;
 	values: Record< string, any >;
 	errors: Record< string, string >;
 	isLoading: boolean;