+++ src/blocks/gravity-form-renderer/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 [View the readme](/src/blocks/gravity-form-renderer/README.md)
 
+## v1.1.0
+
+-   Added: Block supports pre-filled values
+-   Added: Block uses confirmation settings set in Gravity Forms
+
 ## v1.0.0
 
 -   Init!